requirements.txt
apache-airflow
psycopg2-binary
python-dotenv
pandas
numpy
loguru
openpyxl